feature: implemented the file:seek API
The behavior of file:seek API mimics libc's FILE object.

example:

local offset, err = file:seek([whence] [, offset])

The first argument "whence" can be "set", "cur" or "end", each of them
maps to a classical whence:

1) "set" => SEEK_SET
2) "cur" => SEEK_CUR
3) "end" => SEEK_END

The second argument's meaning ("offset") depends on the "whence":

1) If whence is "set", the file offset shall be set to "offset" bytes.
2) If whence is "cur", the file offset shall be set to its current location plus
"offset" ("offset" can be negative).
3) If whence is "end", the file offset shall be set to the size of the file plus
"offset" ("offset" can be negative).

The file, which was opened with the append mode, will always write at
the end of file, this is the inherent behavior of `write` and `writev`.