bugfix: fixed a segmentation fault problem.
The ngx_http_lua_io_file_ctx_metatable_key doesn't hold the io file
object ctx metatable correctly. So the expected __gc method wouldn't be
called while freeing the file_ctx object (an userdata chunk), instead it was
just freed silently. Consequently, our cleanup handler was still hanged on the
r->cleanup linked list (it would be discarded if the __gc method was called),
and it was called when Nginx was freeing current request. The handler, would
try to access the already freed userdata chunk, caused the "use-after-free"
problem, and the process might crash randomly.

See #2 on Github for the details.